LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/cssom-expected.txt

FAIL Simple CSSOM manipulation of subrules The operation is not supported.
FAIL Simple CSSOM manipulation of subrules 1 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('& .broken {}', 3); }" threw object "NotSupportedError: The operation is not supported." that is not a DOMException IndexSizeError: property "code" is equal to 9, expected 1
FAIL Simple CSSOM manipulation of subrules 2 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('& .broken {}', -1); }" threw object "NotSupportedError: The operation is not supported." that is not a DOMException IndexSizeError: property "code" is equal to 9, expected 1
FAIL Simple CSSOM manipulation of subrules 3 assert_throws_dom: function "() => { ss.cssRules[0].deleteRule(5); }" threw object "NotSupportedError: The operation is not supported." that is not a DOMException IndexSizeError: property "code" is equal to 9, expected 1
PASS Simple CSSOM manipulation of subrules 4
FAIL Simple CSSOM manipulation of subrules 5 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('% {}'); }" threw object "NotSupportedError: The operation is not supported." that is not a DOMException SyntaxError: property "code" is equal to 9, expected 12
PASS Simple CSSOM manipulation of subrules 6
FAIL Simple CSSOM manipulation of subrules 7 The operation is not supported.
PASS Simple CSSOM manipulation of subrules 8
FAIL Simple CSSOM manipulation of subrules 9 assert_throws_dom: function "() => { ss.cssRules[0].insertRule('div & {}'); }" threw object "NotSupportedError: The operation is not supported." that is not a DOMException SyntaxError: property "code" is equal to 9, expected 12
FAIL Simple CSSOM manipulation of subrules 10 assert_equals: invalid rule containing ampersand is kept in serialization expected ".a {\n  :is(!& .foo, .b) { color: green; }\n}" but got ".a {\n  :is(.b) { color: green; }\n}"
